# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentinel-X API",
    "description": "Autonomous AI Security Operations Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/config": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Config",
        "operationId": "config_api_v1_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformConfig"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/audit": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "List Audit",
        "operationId": "list_audit_api_v1_audit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Type"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditEntryRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit/verify": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Verify Audit",
        "operationId": "verify_audit_api_v1_audit_verify_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerificationRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AuditEntryRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "chain_index": {
            "type": "integer",
            "title": "Chain Index"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "actor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "actor_type": {
            "type": "string",
            "title": "Actor Type"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "before": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "after": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "After"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "entry_hash": {
            "type": "string",
            "title": "Entry Hash",
            "description": "SHA-256 link in the tamper-evident chain"
          }
        },
        "type": "object",
        "required": [
          "id",
          "chain_index",
          "ts",
          "actor_id",
          "actor_type",
          "action",
          "resource_type",
          "resource_id",
          "before",
          "after",
          "context",
          "correlation_id",
          "entry_hash"
        ],
        "title": "AuditEntryRead"
      },
      "AuditVerificationRead": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "entries_checked": {
            "type": "integer",
            "title": "Entries Checked"
          },
          "head_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Hash"
          },
          "broken_at_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broken At Index"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "entries_checked",
          "head_hash",
          "broken_at_index",
          "reason"
        ],
        "title": "AuditVerificationRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_AuditEntryRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEntryRead"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[AuditEntryRead]"
      },
...
      "PlatformConfig": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "environment": {
            "type": "string",
            "title": "Environment"
          },
          "access_token_ttl_seconds": {
            "type": "integer",
            "title": "Access Token Ttl Seconds"
          },
          "refresh_token_ttl_seconds": {
            "type": "integer",
            "title": "Refresh Token Ttl Seconds"
          },
          "login_rate_limit": {
            "type": "integer",
            "title": "Login Rate Limit"
          },
          "login_rate_window_seconds": {
            "type": "integer",
            "title": "Login Rate Window Seconds"
          },
          "redis_configured": {
            "type": "boolean",
            "title": "Redis Configured"
          },
          "metrics_enabled": {
            "type": "boolean",
            "title": "Metrics Enabled"
          },
          "tracing_enabled": {
            "type": "boolean",
            "title": "Tracing Enabled"
          },
          "jwt_signing_kid": {
            "type": "string",
            "title": "Jwt Signing Kid"
          }
        },
        "type": "object",
        "required": [
          "version",
          "environment",
          "access_token_ttl_seconds",
          "refresh_token_ttl_seconds",
          "login_rate_limit",
          "login_rate_window_seconds",
          "redis_configured",
          "metrics_enabled",
          "tracing_enabled",
          "jwt_signing_kid"
        ],
        "title": "PlatformConfig",
        "description": "Non-secret runtime configuration only."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access JWT from POST /api/v1/auth/login",
        "scheme": "bearer"
      }
    }
  }
}